Organize context providers
