+++ src/components/layout.tsx
@@ -3,19 +3,19 @@ import GlobalStyle from "@style/global-style"
 import Topbar from "@components/topbar/topbar"
 
 import "@style/fonts.css"
-import MediaQueryProvider from "@providers/media-query-provider/media-query-provider"
-import Web3Provider from "@web3/web3-provider"
-import Web3BlocksDataProvider from "@web3/web3-blocks-data-provider"
-import LoadingStatusProvider from "@providers/loading-status-provider/loading-status-provider"
-import ThemeSwitchProvider from "@providers/theme-switch-provider/theme-switch-provider"
+import MediaQueryProvider from "@providers/media-query-provider"
+import Web3Provider from "@providers/web3-provider"
+import Web3BlocksDataProvider from "@providers/web3-blocks-data-provider"
+import LoadingStatusProvider from "@providers/loading-status-provider"
+import ThemeSwitchProvider from "@providers/theme-switch-provider"
 import LoadingIndicator from "@components/loading-indicator/loading-indicator"
-import { USE_MAINNET, BLOCK_COUNT } from "../config"
+import { BLOCK_COUNT } from "../config"
 
 const Layout: React.FunctionComponent = ({ children }) => {
   return (
     <MediaQueryProvider>
       <ThemeSwitchProvider>
-        <Web3Provider useMainnet={USE_MAINNET}>
+        <Web3Provider>
           <Web3BlocksDataProvider maxBlocks={BLOCK_COUNT}>
             <LoadingStatusProvider>
               <>
