+++ src/config.tsx
@@ -3,7 +3,6 @@
  * in a more suitable place.
  */
 export const BLOCK_COUNT = 12 // Total number of recently blocks to display
-export const USE_MAINNET = true // Use mainnet or Rinkeby test net
 export const SUMMARY_MAX_TRANSACTIONS = 60 // Max transactions pip per card
 export const SHUFFLE_MIN_RANGE = 100 // Oldest block you could get when shuffling
 export const SHUFFLE_MAX_RANGE = 7000000 // Newest block you could get when shuffling
